Refactor data types and enhance logging functionality
- Changed `Id` property type from `int` to `long` in multiple classes.
- Removed `Notes` from `DividendDetails` and `Description` from `Result`.
- Added `Pie` and `PieDetail` classes with relevant properties.
- Introduced `PieInstrument` and `PieSettings` classes for better structure.
- Modified `Settings` class to remove unnecessary properties and change `InstrumentShares` to a dictionary.
- Commented out logging lines in `Form1.cs` to adjust logging strategy.
- Updated `GetPieAsync` method to accept `long` for `pieId`.
- Enhanced `TraceOutput` method for better log readability and clarity.
- Implemented `CustomTextWriterTraceListener` for timestamped logging.
- Updated `Program.cs` to utilize the new logging listener with a new log file path.

+++ StockDisplay/Services/ITrading212ApiService.cs
@@ -10,7 +10,7 @@ namespace StockDisplay.Services
 
         // Pies - https://t212public-api-docs.redoc.ly/#tag/Pies
         Task<List<Pie>> GetPiesAsync();
-        Task<PieDetail> GetPieAsync(int pieId); 
+        Task<PieDetail> GetPieAsync(long pieId); 
 
         // Equity Orders - https://t212public-api-docs.redoc.ly/#tag/Equity-Orders
         Task<List<Order>> GetOrdersAsync();
